Fix compilation errors with clang (#1)
* Fix compilation errors with clang

- Remove unused private field
- Change fields to size_t to avoid implicit narrowing
- Change iterator to reference

* Fix examples compilation

- Also add build flags to reference local compilation of PMDK

+++ dcpmm/kvs_dcpmm.h
@@ -25,8 +25,8 @@ struct KVSRef {
   // Specify the encoding type of value content.
   struct KVSHdr hdr;
   // The size of value after encoding.
-  unsigned int size;
-  unsigned int pool_index;
+  size_t size;
+  size_t pool_index;
   // The offset of value on DCPMM in the pool.
   size_t off_in_pool;
 };
